+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixes ``print`` method to mirror builtin
 - Removes default black background for none
 - Adds all effects
+- Objects can be properly added with ``__setattr__``
 
 [v1.0.8](https://github.com/jshwi/object_colors/releases/tag/v1.0.8)  - 2020-02-02
 ------------------------------------------------------------------------